Still no demo commit, hit a roadblock
-The camera building system now works, but it still needs to be tuned.
-Lots of classes moved/deleted/renamed to make everything more clear.
Now most 3D related things are in the 3D package.
-Added SimpleScene, the first implementation of a 3D scene. Contains
background color code as well.
-Added the basics of a Ray Tracer with AA. It is a concept I made that
which should be able to make AA images without (much) extra rays.
-AbstractRenderer now has a purpose.
ROADBLOCK:
A RenderableObject should be able to return polygons or it's shape. But
this turns out to be really tricky to implement in maintainable code.